requirements.txt
openai-whisper
pyaudio
soundfile
speechrecognition
spotipy